web-interface/requirements.txt
streamlit
requests
opencv-python-headless
pandas
pillow